requirements.txt
openai
python-dotenv
aiohttp